refactor: move deprecated code to 'src/auth/deprecated'
We have some deprecated auth code. Not all auth code
is within 'src/auth'. We can eventually look to
consolidate the code in to `src/auth/deprecated`.
