+++ src/extension.ts
@@ -11,7 +11,7 @@ import { activate as activateAwsExplorer } from './awsexplorer/activation'
 import { activate as activateCloudWatchLogs } from './cloudWatchLogs/activation'
 import { initialize as initializeCredentials } from './auth/activation'
 import { initializeAwsCredentialsStatusBarItem } from './auth/statusBarItem'
-import { LoginManager } from './auth/loginManager'
+import { LoginManager } from './auth/deprecated/loginManager'
 import { CredentialsProviderManager } from './auth/providers/credentialsProviderManager'
 import { SharedCredentialsProviderFactory } from './auth/providers/sharedCredentialsProviderFactory'
 import { activate as activateSchemas } from './eventSchemas/activation'
